#443799: git pipeline to build master template (#71)
This Merge Request introduces a new GitHub Actions workflow that builds
the **master template** using the **latest version of the standard
library** and generates a **ZIP archive containing all build
artifacts**, which can be downloaded at the end of the workflow
execution.

At the moment, the workflow is designed to be triggered **manually**. To
run it, you need to:

* Go to **GitHub Actions**
* Select the workflow
* Click **Run workflow**

<img width="470" height="325" alt="image"
src="https://github.com/user-attachments/assets/70c3163c-6036-4177-982a-1702a1023b55"
/>

Please note that the manual trigger **works only on the `main` branch**.

For this reason, and **for testing purposes only**, I also added an
additional trigger:

```yaml
push:                                  # to delete
  branches:                            # to delete
    - 'features/mima/pipeline-to-build-master-template'  # to delete before merging into main, used only for testing
```

This trigger runs the workflow automatically on pushes to the
`features/mima/pipeline-to-build-master-template` branch.
**Before merging into `main`, this trigger must be commented out or
removed**, as it is intended only for development and testing.

---------

Co-authored-by: derB4 <122090401+derB4@users.noreply.github.com>